+++ src/braintree.ios.ts
@@ -1,23 +1,12 @@
 import { Observable } from 'tns-core-modules/data/observable';
 import * as utils from "tns-core-modules/utils/utils";
-import { getAppDelegate } from "./getappdelegate";
+
+const appDelegate = require('./getappdelegate').getAppDelegate();
+const enableMultipleOverridesFor = require('./getappdelegate').enableMultipleOverridesFor;
 
 declare const BTAppSwitch, BTDropInRequest, BTDropInController, UIApplication, PPDataCollector;
 
-export const appDelegate = getAppDelegate();
-
-// enableMultipleOverridesFor method copied from https://github.com/hypery2k/nativescript-urlhandler/blob/45fa7d83d59897db8f3b5077432a154ae542a69e/src/urlhandler.ios.ts#L7
-
-export function enableMultipleOverridesFor(classRef, methodName, nextImplementation) {
-    const currentImplementation = classRef.prototype[methodName];
-    classRef.prototype[methodName] = function () {
-        const result = currentImplementation && currentImplementation.apply(currentImplementation, Array.from(arguments));
-        return nextImplementation.apply(nextImplementation, Array.from(arguments).concat([result]));
-    };
-}
-
 export function setupBraintreeAppDeligate(urlScheme) {
-
     enableMultipleOverridesFor(appDelegate, 'applicationDidFinishLaunchingWithOptions', function (application, launchOptions) {
         try {
             BTAppSwitch.setReturnURLScheme(urlScheme);
