+++ src/getappdelegate.ts
@@ -1,4 +1,5 @@
 import * as application from 'tns-core-modules/application';
+declare const BTAppSwitch;
 
 /**
  * This file has been copied from nativescript-urlhandler
@@ -30,6 +31,17 @@ export function getAppDelegate() {
 
     return application.ios.delegate;
 }
+
+// copied from https://github.com/hypery2k/nativescript-urlhandler/blob/45fa7d83d59897db8f3b5077432a154ae542a69e/src/urlhandler.ios.ts#L7
+
+export function enableMultipleOverridesFor(classRef, methodName, nextImplementation) {
+    const currentImplementation = classRef.prototype[methodName];
+    classRef.prototype[methodName] = function () {
+        const result = currentImplementation && currentImplementation.apply(currentImplementation, Array.from(arguments));
+        return nextImplementation.apply(nextImplementation, Array.from(arguments).concat([result]));
+    };
+}
+
 
 /*
 The MIT License (MIT)
